# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PvP Leaderboard Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List players with filters
      parameters:
        - name: gamemode
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerListResponse"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/{id}/stats:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get player stats by gamemode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStatsResponse"
  /players/{id}/matches:
    get:
      operationId: getPlayerMatches
      tags: [players]
      summary: Get player match history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Player match history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
...
  /submissions:
    post:
      operationId: createSubmission
      tags: [players]
      summary: Submit a match result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
    get:
      operationId: listSubmissions
      tags: [players]
      summary: List submissions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionListResponse"
  /submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [admin]
      summary: Approve a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [admin]
      summary: Reject a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        username:
          type: string
        minecraftUsername:
          type: string
      required:
        - email
        - password
        - username
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [user, moderator, admin, superadmin]
        isBanned:
          type: boolean
        minecraftUsername:
          type: string
        discordId:
          type: string
        discordUsername:
          type: string
        discordAvatar:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - role
        - isBanned
        - createdAt
    Player:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        minecraftUsername:
          type: string
        minecraftUuid:
          type: string
        tier:
          type: string
          enum: [LT1, LT2, LT3, LT4, LT5, HT1, HT2, HT3, HT4, HT5]
        elo:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winStreak:
          type: integer
        discordUsername:
          type: string
        region:
          type: string
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - minecraftUsername
        - tier
        - elo
        - wins
        - losses
        - winStreak
        - createdAt
    PlayerDetail:
      allOf:
        - $ref: "#/components/schemas/Player"
        - type: object
          properties:
            gamemodeStats:
              type: array
              items:
                $ref: "#/components/schemas/GamemodeStat"
            recentMatches:
              type: array
              items:
                $ref: "#/components/schemas/Match"
    PlayerListResponse:
      type: object
      properties:
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - players
        - total
        - page
        - totalPages
    PlayerStatsResponse:
      type: object
      properties:
        gamemodeStats:
          type: array
          items:
            $ref: "#/components/schemas/GamemodeStat"
      required:
        - gamemodeStats
    GamemodeStat:
      type: object
      properties:
        gamemode:
          type: string
          enum: [sword, axe, uhc, vanilla, smp, diapot, nethpot, elytra]
        wins:
          type: integer
        losses:
          type: integer
        elo:
          type: integer
        kd:
          type: number
        winRate:
          type: number
      required:
        - gamemode
        - wins
        - losses
        - elo
        - kd
        - winRate
    Match:
      type: object
      properties:
        id:
          type: integer
        winnerId:
          type: integer
        loserId:
          type: integer
        winnerUsername:
          type: string
        loserUsername:
          type: string
        gamemode:
          type: string
        eloChange:
          type: integer
        playedAt:
          type: string
          format: date-time
      required:
        - id
        - winnerId
        - loserId
        - winnerUsername
        - loserUsername
        - gamemode
        - eloChange
        - playedAt
    MatchListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - matches
        - total
        - page
        - totalPages
...
    SubmissionBody:
      type: object
      properties:
        opponentUsername:
          type: string
        gamemode:
          type: string
        result:
          type: string
          enum: [win, loss]
        evidence:
          type: string
      required:
        - opponentUsername
        - gamemode
        - result
    Submission:
      type: object
      properties:
        id:
          type: integer
        submitterId:
          type: integer
        submitterUsername:
          type: string
        opponentUsername:
          type: string
        gamemode:
          type: string
        result:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        evidence:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - submitterId
        - submitterUsername
        - opponentUsername
        - gamemode
        - result
        - status
        - createdAt
    SubmissionListResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - submissions
        - total
        - page
        - totalPages
...